src/main/java/dev/voidbit/lapiz/StudentController.java
package dev.voidbit.lapiz;

import org.springframework.data.annotation.Id;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

import javax.persistence.GeneratedValue;

@RestController
public class StudentController {
    private StudentService studentService;

    public StudentController(StudentService studentService){
        this.studentService = studentService;
    }

    @GetMapping("/student/{id}")
    private Student getStudentById(@PathVariable Long id){
        return studentService.getStudentById(id);
    }
}
